+++ customers/ms-customers/ms-customers/src/main/java/igraciarena/poc/mscustomers/application/exceptions/NotFoundException.java
@@ -1,4 +1,4 @@
-package igraciarena.poc.mscustomers.exceptions;
+package igraciarena.poc.mscustomers.application.exceptions;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
