+++ customers/ms-customers/ms-customers/src/main/java/igraciarena/poc/mscustomers/application/input/CreateCustomerUseCase.java
@@ -1,0 +1,9 @@
+package igraciarena.poc.mscustomers.application.input;
+
+
+import igraciarena.poc.mscustomers.application.input.commands.CreateCustomerCommand;
+
+public interface CreateCustomerUseCase {
+
+    void createCustomer(CreateCustomerCommand createCustomerCommand);
+}
