+++ customers/ms-customers/ms-customers/src/main/java/igraciarena/poc/mscustomers/application/input/CreateRegionUseCase.java
@@ -1,0 +1,9 @@
+package igraciarena.poc.mscustomers.application.input;
+
+
+import igraciarena.poc.mscustomers.application.input.commands.CreateRegionCommand;
+
+public interface CreateRegionUseCase {
+
+    void createRegion(CreateRegionCommand createRegionCommand);
+}
