+++ customers/ms-customers/ms-customers/src/main/java/igraciarena/poc/mscustomers/application/input/DeleteCustomerUseCase.java
@@ -1,0 +1,9 @@
+package igraciarena.poc.mscustomers.application.input;
+
+
+import igraciarena.poc.mscustomers.application.input.commands.DeleteCustomerCommand;
+
+public interface DeleteCustomerUseCase {
+
+    void deleteCustomer(DeleteCustomerCommand deleteCustomerCommand);
+}
