+++ customers/ms-customers/ms-customers/src/main/java/igraciarena/poc/mscustomers/application/input/DeleteRegionUseCase.java
@@ -1,0 +1,9 @@
+package igraciarena.poc.mscustomers.application.input;
+
+
+import igraciarena.poc.mscustomers.application.input.commands.DeleteRegionCommand;
+
+public interface DeleteRegionUseCase {
+
+    void deleteRegion(DeleteRegionCommand deleteRegionCommand);
+}
