+++ customers/ms-customers/ms-customers/src/main/java/igraciarena/poc/mscustomers/application/input/UpdateCustomerUseCase.java
@@ -1,0 +1,9 @@
+package igraciarena.poc.mscustomers.application.input;
+
+
+import igraciarena.poc.mscustomers.application.input.commands.UpdateCustomerCommand;
+
+public interface UpdateCustomerUseCase {
+
+    void updateCustomer(UpdateCustomerCommand updateCustomerCommand);
+}
