+++ customers/ms-customers/ms-customers/src/main/java/igraciarena/poc/mscustomers/application/input/UpdateRegionUseCase.java
@@ -1,0 +1,9 @@
+package igraciarena.poc.mscustomers.application.input;
+
+
+import igraciarena.poc.mscustomers.application.input.commands.UpdateRegionCommand;
+
+public interface UpdateRegionUseCase {
+
+    void updateRegionUseCase(UpdateRegionCommand updateRegionCommand);
+}
