+++ customers/ms-customers/ms-customers/src/main/java/igraciarena/poc/mscustomers/application/input/commands/DeleteCustomerCommand.java
@@ -1,11 +1,11 @@
-package igraciarena.poc.mscustomers.input.querys;
+package igraciarena.poc.mscustomers.application.input.commands;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
 @Getter
 @AllArgsConstructor
-public class FindCustomerByIdQuery {
+public class DeleteCustomerCommand {
 
     private final Long customerId;
 }
