+++ customers/ms-customers/ms-customers/src/main/java/igraciarena/poc/mscustomers/application/input/commands/IncreaseNumberInvoicesCommand.java
@@ -1,10 +1,11 @@
-package igraciarena.poc.mscustomers.input.querys;
+package igraciarena.poc.mscustomers.application.input.commands;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
 @Getter
 @AllArgsConstructor
-public class FindInvoicesByCustomerIdQuery {
+public class IncreaseNumberInvoicesCommand {
+
     private final Long customerId;
 }
