+++ customers/ms-customers/ms-customers/src/main/java/igraciarena/poc/mscustomers/application/input/commands/UpdateRegionCommand.java
@@ -1,4 +1,4 @@
-package igraciarena.poc.mscustomers.input.commands;
+package igraciarena.poc.mscustomers.application.input.commands;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
