+++ customers/ms-customers/ms-customers/src/main/java/igraciarena/poc/mscustomers/application/input/querys/FindCustomerByIdQuery.java
@@ -1,11 +1,11 @@
-package igraciarena.poc.mscustomers.input.commands;
+package igraciarena.poc.mscustomers.application.input.querys;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
 @Getter
 @AllArgsConstructor
-public class IncreaseNumberInvoicesCommand {
+public class FindCustomerByIdQuery {
 
     private final Long customerId;
 }
