+++ customers/ms-customers/ms-customers/src/main/java/igraciarena/poc/mscustomers/application/input/querys/FindInvoicesByCustomerIdQuery.java
@@ -1,11 +1,10 @@
-package igraciarena.poc.mscustomers.input.commands;
+package igraciarena.poc.mscustomers.application.input.querys;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
 @Getter
 @AllArgsConstructor
-public class DeleteCustomerCommand {
-
+public class FindInvoicesByCustomerIdQuery {
     private final Long customerId;
 }
