+++ customers/ms-customers/ms-customers/src/main/java/igraciarena/poc/mscustomers/application/output/CreateRegionService.java
@@ -1,8 +1,8 @@
 package igraciarena.poc.mscustomers.application.output;
 
 import igraciarena.poc.mscustomers.adapters.output.persistance.RegionRepository;
-import igraciarena.poc.mscustomers.input.CreateRegionUseCase;
-import igraciarena.poc.mscustomers.input.commands.CreateRegionCommand;
+import igraciarena.poc.mscustomers.application.input.CreateRegionUseCase;
+import igraciarena.poc.mscustomers.application.input.commands.CreateRegionCommand;
 import igraciarena.poc.mscustomers.domain.mappers.RegionMapper;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
