+++ customers/ms-customers/ms-customers/src/main/java/igraciarena/poc/mscustomers/application/output/FindCustomerByIdService.java
@@ -3,9 +3,9 @@ package igraciarena.poc.mscustomers.application.output;
 
 import igraciarena.poc.mscustomers.adapters.output.dbentities.Customer;
 import igraciarena.poc.mscustomers.adapters.output.persistance.CustomerRepository;
-import igraciarena.poc.mscustomers.exceptions.NotFoundException;
-import igraciarena.poc.mscustomers.input.FindCustomerByIdUseCase;
-import igraciarena.poc.mscustomers.input.querys.FindCustomerByIdQuery;
+import igraciarena.poc.mscustomers.application.exceptions.NotFoundException;
+import igraciarena.poc.mscustomers.application.input.FindCustomerByIdUseCase;
+import igraciarena.poc.mscustomers.application.input.querys.FindCustomerByIdQuery;
 import igraciarena.poc.mscustomers.domain.dtos.CustomerResponse;
 import igraciarena.poc.mscustomers.domain.mappers.CustomerMapper;
 import lombok.AllArgsConstructor;
