+++ customers/ms-customers/ms-customers/src/main/java/igraciarena/poc/mscustomers/application/output/IncreaseNumberInvoicesService.java
@@ -2,9 +2,9 @@ package igraciarena.poc.mscustomers.application.output;
 
 import igraciarena.poc.mscustomers.adapters.output.dbentities.Customer;
 import igraciarena.poc.mscustomers.adapters.output.persistance.CustomerRepository;
-import igraciarena.poc.mscustomers.exceptions.NotFoundException;
-import igraciarena.poc.mscustomers.input.IncreaseNumberInvoicesUseCase;
-import igraciarena.poc.mscustomers.input.commands.IncreaseNumberInvoicesCommand;
+import igraciarena.poc.mscustomers.application.exceptions.NotFoundException;
+import igraciarena.poc.mscustomers.application.input.IncreaseNumberInvoicesUseCase;
+import igraciarena.poc.mscustomers.application.input.commands.IncreaseNumberInvoicesCommand;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
