+++ customers/ms-customers/ms-customers/src/main/java/igraciarena/poc/mscustomers/application/output/UpdateCustomerService.java
@@ -4,9 +4,9 @@ import igraciarena.poc.mscustomers.adapters.output.dbentities.Customer;
 import igraciarena.poc.mscustomers.adapters.output.dbentities.Region;
 import igraciarena.poc.mscustomers.adapters.output.persistance.CustomerRepository;
 import igraciarena.poc.mscustomers.adapters.output.persistance.RegionRepository;
-import igraciarena.poc.mscustomers.exceptions.NotFoundException;
-import igraciarena.poc.mscustomers.input.UpdateCustomerUseCase;
-import igraciarena.poc.mscustomers.input.commands.UpdateCustomerCommand;
+import igraciarena.poc.mscustomers.application.exceptions.NotFoundException;
+import igraciarena.poc.mscustomers.application.input.UpdateCustomerUseCase;
+import igraciarena.poc.mscustomers.application.input.commands.UpdateCustomerCommand;
 import igraciarena.poc.mscustomers.domain.mappers.CustomerMapper;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
