--- customers/ms-customers/ms-customers/src/main/java/igraciarena/poc/mscustomers/input/CreateCustomerUseCase.java
@@ -1,9 +1,0 @@
-package igraciarena.poc.mscustomers.input;
-
-
-import igraciarena.poc.mscustomers.input.commands.CreateCustomerCommand;
-
-public interface CreateCustomerUseCase {
-
-    void createCustomer(CreateCustomerCommand createCustomerCommand);
-}
