--- customers/ms-customers/ms-customers/src/main/java/igraciarena/poc/mscustomers/input/CreateRegionUseCase.java
@@ -1,9 +1,0 @@
-package igraciarena.poc.mscustomers.input;
-
-
-import igraciarena.poc.mscustomers.input.commands.CreateRegionCommand;
-
-public interface CreateRegionUseCase {
-
-    void createRegion(CreateRegionCommand createRegionCommand);
-}
