--- customers/ms-customers/ms-customers/src/main/java/igraciarena/poc/mscustomers/input/DeleteCustomerUseCase.java
@@ -1,9 +1,0 @@
-package igraciarena.poc.mscustomers.input;
-
-
-import igraciarena.poc.mscustomers.input.commands.DeleteCustomerCommand;
-
-public interface DeleteCustomerUseCase {
-
-    void deleteCustomer(DeleteCustomerCommand deleteCustomerCommand);
-}
