--- customers/ms-customers/ms-customers/src/main/java/igraciarena/poc/mscustomers/input/DeleteRegionUseCase.java
@@ -1,9 +1,0 @@
-package igraciarena.poc.mscustomers.input;
-
-
-import igraciarena.poc.mscustomers.input.commands.DeleteRegionCommand;
-
-public interface DeleteRegionUseCase {
-
-    void deleteRegion(DeleteRegionCommand deleteRegionCommand);
-}
