--- customers/ms-customers/ms-customers/src/main/java/igraciarena/poc/mscustomers/input/UpdateCustomerUseCase.java
@@ -1,9 +1,0 @@
-package igraciarena.poc.mscustomers.input;
-
-
-import igraciarena.poc.mscustomers.input.commands.UpdateCustomerCommand;
-
-public interface UpdateCustomerUseCase {
-
-    void updateCustomer(UpdateCustomerCommand updateCustomerCommand);
-}
