--- customers/ms-customers/ms-customers/src/main/java/igraciarena/poc/mscustomers/input/UpdateRegionUseCase.java
@@ -1,9 +1,0 @@
-package igraciarena.poc.mscustomers.input;
-
-
-import igraciarena.poc.mscustomers.input.commands.UpdateRegionCommand;
-
-public interface UpdateRegionUseCase {
-
-    void updateRegionUseCase(UpdateRegionCommand updateRegionCommand);
-}
