+++ customers/ms-customers/ms-customers/src/test/java/igraciarena/poc/mscustomers/application/output/CreateRegionTestStub.java
@@ -2,7 +2,7 @@ package igraciarena.poc.mscustomers.application.output;
 
 import igraciarena.poc.mscustomers.adapters.output.dbentities.Region;
 import igraciarena.poc.mscustomers.adapters.output.persistance.RegionRepository;
-import igraciarena.poc.mscustomers.input.commands.CreateRegionCommand;
+import igraciarena.poc.mscustomers.application.input.commands.CreateRegionCommand;
 import igraciarena.poc.mscustomers.utils.StubFactory;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
