+++ customers/ms-customers/ms-customers/src/test/java/igraciarena/poc/mscustomers/application/output/FindCustomerByIdServiceTest.java
@@ -3,7 +3,7 @@ package igraciarena.poc.mscustomers.application.output;
 import igraciarena.poc.mscustomers.adapters.output.dbentities.Customer;
 import igraciarena.poc.mscustomers.adapters.output.persistance.CustomerRepository;
 import igraciarena.poc.mscustomers.domain.dtos.CustomerResponse;
-import igraciarena.poc.mscustomers.input.querys.FindCustomerByIdQuery;
+import igraciarena.poc.mscustomers.application.input.querys.FindCustomerByIdQuery;
 import igraciarena.poc.mscustomers.utils.StubFactory;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
