+++ customers/ms-customers/ms-customers/src/test/java/igraciarena/poc/mscustomers/application/output/FindRegionByIdServiceTest.java
@@ -3,7 +3,7 @@ package igraciarena.poc.mscustomers.application.output;
 import igraciarena.poc.mscustomers.adapters.output.dbentities.Region;
 import igraciarena.poc.mscustomers.adapters.output.persistance.RegionRepository;
 import igraciarena.poc.mscustomers.domain.dtos.RegionResponse;
-import igraciarena.poc.mscustomers.input.querys.FindRegionByIdQuery;
+import igraciarena.poc.mscustomers.application.input.querys.FindRegionByIdQuery;
 import igraciarena.poc.mscustomers.utils.StubFactory;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
