+++ customers/ms-customers/ms-customers/src/test/java/igraciarena/poc/mscustomers/application/output/UpdateCustomerServiceTest.java
@@ -1,10 +1,9 @@
 package igraciarena.poc.mscustomers.application.output;
 
 import igraciarena.poc.mscustomers.adapters.output.dbentities.Customer;
-import igraciarena.poc.mscustomers.adapters.output.dbentities.Region;
 import igraciarena.poc.mscustomers.adapters.output.persistance.CustomerRepository;
 import igraciarena.poc.mscustomers.adapters.output.persistance.RegionRepository;
-import igraciarena.poc.mscustomers.input.commands.UpdateCustomerCommand;
+import igraciarena.poc.mscustomers.application.input.commands.UpdateCustomerCommand;
 import igraciarena.poc.mscustomers.utils.StubFactory;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
